requirements.txt
mysql-connector-python==8.0.30
pandas==1.4.0
pyflowchart==0.2.3
selenium==4.4.2
numpy~=1.22.2